requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Session>=0.5.0
flasgger>=0.9.7.1
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
gunicorn>=21.2.0
pytest>=7.0.0
pytest-cov>=4.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
